reorder documents table
