+++ packages/client/src/components/charges/charge-extended-info.tsx
@@ -18,11 +18,11 @@ import {
 } from '../../gql/graphql.js';
 import { FragmentType, isFragmentReady } from '../../gql/index.js';
 import { BusinessTripSummarizedReport, RegenerateLedgerRecordsButton } from '../common/index.js';
+import { DocumentsGallery } from '../documents-table/documents-gallery.js';
+import { DocumentsTable } from '../documents-table/index.js';
 import { LedgerTable } from '../ledger-table/index.js';
 import { ChargeErrors } from './charge-errors.jsx';
 import { ChargeTransactionsTable } from './charge-transactions-table.jsx';
-import { DocumentsGallery } from './documents/documents-gallery.jsx';
-import { DocumentsTable } from './documents/documents-table.jsx';
 import { ConversionInfo } from './extended-info/conversion-info.jsx';
 import { CreditcardTransactionsInfo } from './extended-info/creditcard-transactions-info.jsx';
 import { ChargeMiscExpensesTable } from './extended-info/misc-expenses.jsx';
