+++ packages/client/src/components/documents-table/cells/serial.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
 import { Indicator } from '@mantine/core';
-import { DocumentSerialFieldsFragmentDoc, DocumentType } from '../../../../gql/graphql.js';
-import { FragmentType, getFragmentData } from '../../../../gql/index.js';
+import { DocumentSerialFieldsFragmentDoc, DocumentType } from '../../../gql/graphql.js';
+import { FragmentType, getFragmentData } from '../../../gql/index.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
 /* GraphQL */ `
