+++ packages/client/src/components/documents-table/documents-table-row.tsx
@@ -4,9 +4,9 @@ import {
   DocumentTableRowDocument,
   TableDocumentsRowFieldsFragment,
   TableDocumentsRowFieldsFragmentDoc,
-} from '../../../gql/graphql.js';
-import { FragmentType, getFragmentData } from '../../../gql/index.js';
-import { EditMiniButton } from '../../common/index.js';
+} from '../../gql/graphql.js';
+import { FragmentType, getFragmentData } from '../../gql/index.js';
+import { EditMiniButton } from '../common/index.js';
 import { Amount, Creditor, DateCell, Debtor, Files, Serial, TypeCell, Vat } from './cells/index.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
